+++ tests/unit/test_dimer_scan_cli.py
@@ -13,6 +13,7 @@ from mmml.cli.registry import command_by_name
 EXPECTED_CALCULATORS = {
     "physnet",
     "kernnn",
+    "metatomic",
     "spookynet",
     "mbd",
     "multipoles",